requirements.txt
streamlit
opencv-python
moviepy==1.0.3
moviepy
faster-whisper
groq
python-dotenv
numpy